Generalize metric-based assertion class and add mapping class
